ilearning_ics_kit/requirements.txt
requests
beautifulsoup4
lxml
ics
python-dateutil
pytz
